Make first move to numba montecarlo tests
